fixes sintax
