+++ backend/controllers/all.js
@@ -1,0 +1,7 @@
+const { POKENEAS } = require('../mocks/pokeneas');
+
+const getAllPokenea = () => {
+    return POKENEAS;
+}
+
+module.exports = { getAllPokenea };
