+++ frontend/.env
@@ -1,1 +1,1 @@
-VITE_BACKEND_URL=http://localhost:3001/
+VITE_BACKEND_URL=http://localhost:3001
