+++ frontend/src/services/fetchPokeneas.js
@@ -1,17 +1,17 @@
 export async function fetchFullPokenea() {
-    const response = await fetch(`${import.meta.env.VITE_BACKEND_URL}api/full`);
+    const response = await fetch(`${import.meta.env.VITE_BACKEND_URL}/api/full`);
     const data = await response.json(); 
     return data;
 }
 
 export async function fetchRawPokenea() {
-    const response = await fetch(`${import.meta.env.VITE_BACKEND_URL}api/raw`);
+    const response = await fetch(`${import.meta.env.VITE_BACKEND_URL}/api/raw`);
     const data = await response.json(); 
     return data;
 }
 
 export async function fetchAllPokeneas() {
-    const response = await fetch(`${import.meta.env.VITE_BACKEND_URL}api/all`);
+    const response = await fetch(`${import.meta.env.VITE_BACKEND_URL}/api/all`);
     const data = await response.json(); 
     return data;
 }
