+++ frontend/vite.config.js
@@ -3,7 +3,7 @@ import { defineConfig } from 'vite';
 export default defineConfig({
   server: {
     proxy: {
-      '/api': process.env.VITE_BACKEND_URL,
+      '/api': process.env.VITE_BACKEND_URL || 'http://localhost:3001',
     },
   },
 });
